Add callback to fs.unlink to remove depreciation and log any errors

+++ api/upload-service.js
@@ -6,28 +6,31 @@ const mime = require('mime-types');
 const path = require('path');
 const Grid = require('gridfs-stream');
 
-Grid.mongo = db.mongo;
-
-exports.uploadImage = (pathFile, fileName, mimetype) => {
+exports.uploadImage = (pathFile, fileName, mimetype, willUnlink = true) => {
   return new Promise((resolve, reject) => {
-    const gfs = Grid(db.conn.db);
+    const gfs = Grid(db.conn.db, db.mongo);
+
     const writestream = gfs.createWriteStream({
       filename: fileName,
       content_type: mimetype,
     });
     fs.createReadStream(pathFile).pipe(writestream);
     writestream.on('close', (file) => {
-      fs.unlink(pathFile, ((err) => {
-        if (err) return reject({ code: 1, message: err });
+      if (willUnlink) {
+        fs.unlink(pathFile, ((err) => {
+          if (err) return reject({ code: 1, message: err });
+          resolve(file._id);
+        }));
+      } else {
         resolve(file._id);
-      }));
+      }
     });
   });
 };
 
 exports.downloadImage = (idImage) => {
   return new Promise((resolve, reject) => {
-    const gfs = Grid(db.conn.db);
+    const gfs = Grid(db.conn.db, db.mongo);
 
     gfs.files.find({ _id: idImage }).toArray((err, files) => {
       if (files.length === 0 || err) return reject({ code: 1, message: err });
@@ -48,7 +51,7 @@ exports.downloadImage = (idImage) => {
 
 exports.findDefaultAvatarId = (imageName, hash) => {
   return new Promise((resolve, reject) => {
-    const gfs = Grid(db.conn.db);
+    const gfs = Grid(db.conn.db, db.mongo);
 
     gfs.files.find({ filename: imageName, md5: hash }).toArray((err, files) => {
       if (files.length === 0 || err) return reject({ code: 1, message: err });
@@ -60,7 +63,7 @@ exports.findDefaultAvatarId = (imageName, hash) => {
 
 exports.deleteImage = (idImage) => {
   return new Promise((resolve, reject) => {
-    const gfs = Grid(db.conn.db);
+    const gfs = Grid(db.conn.db, db.mongo);
 
     gfs.remove({
       _id: idImage,
